* appdefault.c: Fix constness to avoid warning. * init_ctx.c: Do the same stuff on the Mac as on Unix. * preauth2.c: Added cast to fix warning
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@15258 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/krb/preauth2.c
@@ -737,7 +737,7 @@ krb5_error_code pa_sam_2(krb5_context context,
    /* enc_sam_response_enc_2 from above */
 
    retval = krb5_c_encrypt_length(context, as_key->enctype, scratch->length,
-		(unsigned int *) &sr2.sam_enc_nonce_or_sad.ciphertext.length);
+		(size_t *) &sr2.sam_enc_nonce_or_sad.ciphertext.length);
    if (retval) {
 	krb5_free_sam_challenge_2(context, sc2);
 	krb5_free_sam_challenge_2_body(context, sc2b);
